remove remaining psi reports
