+++ corehq/__init__.py
@@ -4,7 +4,6 @@ from corehq.apps.reports.standard import (monitoring, inspect, export,
     deployments, sms, ivr)
 import corehq.apps.receiverwrapper.reports as receiverwrapper
 import phonelog.reports as phonelog
-from corehq.apps.reports.commtrack import psi_prototype
 from corehq.apps.reports.commtrack import standard as commtrack_reports
 from corehq.apps.reports.commtrack import maps as commtrack_maps
 from corehq.apps.reports.commconnect import system_overview
@@ -49,9 +48,6 @@ def REPORTS(project):
             commtrack_reports.CurrentStockStatusReport,
             commtrack_reports.AggregateStockStatusReport,
             commtrack_reports.RequisitionReport,
-            psi_prototype.VisitReport,
-            psi_prototype.StockOutReport,
-            psi_prototype.StockReportExport,
             commtrack_maps.StockStatusMapReport,
             commtrack_maps.ReportingStatusMapReport,
         )))
