+++ corehq/apps/reports/commtrack/maps.py
@@ -1,7 +1,7 @@
-from corehq.apps.reports.commtrack.psi_prototype import CommtrackReportMixin
 from django.utils.translation import ugettext_noop
 from corehq.apps.commtrack.models import Product
 from django.template.loader import render_to_string
+from corehq.apps.reports.commtrack.standard import CommtrackReportMixin
 from corehq.apps.reports.standard.maps import GenericMapReport
